+++ src/embeddings.jl
@@ -89,8 +89,7 @@ Reconstruct `s` using the delay coordinates embedding with `γ` temporal neighbo
 and delay `τ` and return the result as a [`Dataset`](@ref). Optionally use weight `w`.
 
 Use [`embed`](@ref) for the version that accepts the embedding dimension `D = γ+1`
-instead. Here `τ` is strictly positive, use [`genembed`](@ref) for a generalized
-version.
+instead. Here `τ ≥ 0`, use [`genembed`](@ref) for a generalized version.
 
 ## Description
 ### Single Timeseries
@@ -121,10 +120,7 @@ of the embedded vector are further weighted with ``w^\\gamma``, like so
 
 ### Multiple Timeseries
 To make a reconstruction out of a multiple timeseries (i.e. trajectory) the number
-of timeseries must be known by type, so `s` can be either:
-
-* `s::AbstractDataset{B}`
-* `s::SizedAray{A, B}`
+of timeseries must be known by type, so `s` must be a `Dataset`.
 
 If the trajectory is for example ``(x, y)`` and `τ` is integer, then the ``n``-th
 entry of the embedded space is
@@ -195,14 +191,14 @@ embed(s, D, τ) = reconstruct(s, D-1, τ)
 """
     MTDelayEmbedding(γ, τ, B) -> `embedding`
 Return a delay coordinates embedding structure to be used as a functor,
-given multiple timeseries (`B` in total), either as a [`Dataset`](@ref) or a
-`SizedArray`, and some index.
+that embeds multiple timeseries (`B` in total) given in the form of a [`Dataset`](@ref).
+
 Calling
 ```julia
 embedding(s, n)
 ```
-will create the `n`-th delay vector of the embedded space, which has `γ`
-temporal neighbors with delay(s) `τ`. See [`reconstruct`](@ref) for more.
+where `s` is a `Dataset` will create the `n`-th delay vector of the embedded space,
+which has `γ` temporal neighbors with delay(s) `τ`. See [`reconstruct`](@ref) for more.
 
 **Be very careful when choosing `n`, because `@inbounds` is used internally.
 It must be that `n ≤ length(s) - maximum(τ)`.**
@@ -238,6 +234,7 @@ end
 @generated function (r::MTDelayEmbedding{γ, B, X})(
     s::Union{AbstractDataset{B, T}, SizedArray{Tuple{A, B}, T, 2, M}},
     i) where {γ, A, B, T, M, X}
+    typeof(s) <: SizedArray && @warn "Using SizedArrays is deprecated. Use Dataset instead."
     gensprev = [:(s[i, $d]) for d=1:B]
     gens = [:(s[i + r.delays[$k, $d], $d]) for k=1:γ for d=1:B]
     quote
@@ -245,6 +242,8 @@ end
         @inbounds return SVector{$(γ+1)*$B,T}($(gensprev...), $(gens...))
     end
 end
+
+τrange(s, de::MTDelayEmbedding) = 1:(size(s)[1] - maximum(de.delays))
 
 @inline function reconstruct(
     s::Union{AbstractDataset{B, T}, SizedArray{Tuple{A, B}, T, 2, M}},